+++ packages/core/src/index.ts
@@ -11,6 +11,5 @@ export * from './utils/method';
 export * from './utils/httpStatusCode';
 export * from './utils/isCrushed';
 export * from './utils/actionType';
-export * from './utils/keepRegister';
 export * from './services/BaseHttpService';
 export * from './services/HttpServiceBuilder';