+++ packages/react/tests/models/KeepRegisterModel.spec.ts
@@ -8,9 +8,9 @@ beforeEach(() => {
 });
 
 it ('can keep register by sub model', () => {
-  model = new KeepRegisterModel(100, Math.random().toString());
+  model = KeepRegisterModel.init(100, Math.random().toString());
   expect(model.data.count).toBe(100);
 
-  model = new KeepRegisterModel(55, Math.random().toString());
+  model = KeepRegisterModel.init(55, Math.random().toString());
   expect(model.data.count).toBe(55);
 });