+++ packages/vue/tests/models/KeepRegisterModel.ts
@@ -1,4 +1,3 @@
-import { keepRegister } from '@redux-model/core';
 import { Model } from '../../src';
 
 interface Data {
@@ -9,11 +8,11 @@ export class KeepRegisterModel extends Model<Data> {
   protected readonly stateCount: number;
 
   constructor(count: number = 0, alias: string = '') {
-    const register = keepRegister(KeepRegisterModel);
     super(alias);
     this.stateCount = count;
-    register();
   }
+
+  test = this.action(() => {});
 
   protected initialState(): Data {
     return {